Fix addModlist on python3.
With python3 filter returns a filter object, which will always be true.
Use a more-pythonic list comprehension instead of filter.

Update other usage of this pattern to not use filter either.

Note that this already has a test (test_modlist.py) that picked this up,
but its not a standard python unittest so isn't run by default.